moved structs and enums to a separate Structs.hpp
